Move automatic ship protection to claims
Allows ship owners to add or remove builders from their ship.
Retroactively applied.
